% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXKCD.R
\name{saveConfig}
\alias{saveConfig}
\title{Save XKCD database info into a file in the user directory}
\usage{
saveConfig()
}
\description{
This function saves the xkcd database as a file in the user's home
directory.
}
\references{
http://xkcd.com/license.html
}

