% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXKCD.R
\name{searchXKCD}
\alias{searchXKCD}
\title{Search your favorite XKCD comic strip by title/transcript}
\usage{
searchXKCD(which = "significant")
}
\arguments{
\item{which}{string.}
}
\value{
a data.frame containing the following fields:
\itemize{
    \item num The num of the XKCD comic strip
    \item title The title of the XKCD comic strip
}
}
\description{
This function uses grep to inspect the title and transcript for all
the occurrences of a specified string and returns a data.frame with
both the number and the title of the XKCD comic strips.
}
\examples{
\dontrun{
library("RXKCD")
searchXKCD(which="significant")
searchXKCD(which="someone is wrong") }

}
\references{
\url{http://xkcd.com/license.html}
}
