% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXKCD.R
\name{updateConfig}
\alias{updateConfig}
\title{Update the XKCD database saved in the user directory}
\usage{
updateConfig()
}
\description{
This function updates the local version of the XKCD database used
by searchXKCD().
}
\references{
\url{https://xkcd.com/license.html}
}
